re-added table contents to all schemas, remoduled models.rs to only have Crud trait, added users.rs with all the Crud implementations
