+++ src/database.rs
@@ -1,2 +1,3 @@
 pub mod models;
 pub mod connection;
+pub mod users;
